web_server/requirements.txt
Flask==1.1.2
Pillow==7.1.2
requests==2.23.0
tensorflow==2.2.0
numpy==1.18.4
gunicorn==20.0.4
redis
